Change uint to uint256

+++ contracts/hotel/Hotel.sol
@@ -225,8 +225,8 @@ contract Hotel is PrivateCall, Images {
   function book(
     address unitAddress,
     address from,
-    uint fromDay,
-    uint daysAmount
+    uint256 fromDay,
+    uint256 daysAmount
   ) fromSelf() {
     require(unitsIndex[unitAddress] > 0);
     require(daysAmount > 0);
@@ -244,8 +244,8 @@ contract Hotel is PrivateCall, Images {
   function bookWithLif(
     address unitAddress,
     address from,
-    uint fromDay,
-    uint daysAmount
+    uint256 fromDay,
+    uint256 daysAmount
   ) fromSelf() {
     require(unitsIndex[unitAddress] > 0);
     require(daysAmount > 0);
